+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-## [1.1.2] - 2023-11-01
+## [1.2.0] - 2023-11-05
+
+- Dark mode support
+- Added header divider (configurable)
+- Added footer section with divider (configurable)
+- Tailwind flavor now uses data attributes to style elements
+- Updated look and feel
+- Simplified code a bit
+
+## [1.1.3] - 2023-11-01
 
 - Added configuration block